+++ R/fw13_loadMeta.R
@@ -55,7 +55,7 @@ fw13_loadMeta <- function(
     if ( verbose ) {
       if ( !forceDownload )
         message("Could not find local data.")
-      message(paste("Downloading and saving data to", filePath, "."))
+      message(paste("Downloading and saving data to", filePath))
     }
     
     # If local data does not exist, download and return it.