+++ R/wrcc_loadMeta.R
@@ -60,7 +60,7 @@ wrcc_loadMeta <- function(
     if ( verbose ) {
       if ( !forceDownload )
         message("Could not find local data.")
-      message(paste("Downloading and saving data to", filePath, "."))
+      message(paste("Downloading and saving data to", filePath))
     }
     
     # If local data does not exist, download and return it.